+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/app/App.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
 import com.adammcneilly.spacenerd.scaffold.LocalSharedTransitionScope
-import com.adammcneilly.spacenerd.shared.theme.SpaceTheme
+import com.adammcneilly.spacenerd.shared.ui.theme.SpaceTheme
 
 @Composable
 @OptIn(ExperimentalSharedTransitionApi::class)
