+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/navigation/HomeTab.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.spacenerd.shared
+package com.adammcneilly.spacenerd.shared.navigation
 
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Groups
