+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/navigation/NavItem.kt
@@ -1,6 +1,6 @@
 package com.adammcneilly.spacenerd.shared.navigation
 
-import com.adammcneilly.spacenerd.shared.HomeTab
+import com.adammcneilly.spacenerd.shared.navigation.HomeTab
 
 data class NavItem(
     val tab: HomeTab,
