+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/scaffold/ScaffoldState.kt
