+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.spacenerd.shared.theme
+package com.adammcneilly.spacenerd.shared.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
