+++ shared/src/commonMain/kotlin/com/adammcneilly/spacenerd/shared/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.adammcneilly.spacenerd.shared.theme
+package com.adammcneilly.spacenerd.shared.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
